requirements.txt
flask
mysql-connector-python
flask-cors
flasgger
python-dotenv
fastapi
uvicorn[standard]
sqlalchemy
python-jose
python-multipart
passlib[bcrypt]==1.7.4
bcrypt==4.0.1